+++ Multiplayer Bullet Hell/Assets/Scripts/PlayerController.cs
@@ -8,43 +8,31 @@ public class PlayerController : NetworkBehaviour
 {
     #region fields
     public float walkingSpeed = 4.0f;
-    public float gravity = 20.0f;
     private CharacterController characterController;
     private Vector3 moveDirection = Vector3.zero;
 
     private Camera cam;
 
-    private List<GameObject> bulletPool = new List<GameObject>();
-    public GameObject bullet;
-    private int currentBulletIndex = 0;
-
-    private int lifeTotal = 1;
     private bool isGameOver = false;
+
+    // Synced variables
+    [SyncVar]
+    private int lifeTotal = 1;
     #endregion
 
 
     void Start()
     {
-        for(int i = 0; i < 10; i++)
-        {
-            bulletPool.Add(Instantiate(bullet));
-            //bulletPool[i].SetActive(false);
-        }
         characterController = GetComponent<CharacterController>();
         cam = Camera.main;
     }
 
-    
+    [ClientCallback]
     void Update()
     {
         if(isLocalPlayer)
         {
-            Move();
-
-            if (Input.GetKeyDown(KeyCode.Mouse0))
-            {
-                CmdShoot();
-            }
+            CmdMove(CalculateMove());
 
             if(lifeTotal <= 0)
             {
@@ -55,11 +43,12 @@ public class PlayerController : NetworkBehaviour
         }
     }
 
-    #region helper methods
+    #region Helper Methods
     /// <summary>
-    /// Moves the player through inputs.
+    /// Moves the player using inputs.
     /// </summary>
-    private void Move()
+    [Client]
+    public Vector3 CalculateMove()
     {
         Vector3 forward = transform.TransformDirection(Vector3.forward);
         Vector3 right = transform.TransformDirection(Vector3.right);
@@ -68,45 +57,22 @@ public class PlayerController : NetworkBehaviour
         float curSpeedY = walkingSpeed * Input.GetAxis("Horizontal");
         moveDirection = (forward * curSpeedX) + (right * curSpeedY);
 
-        // Apply gravity
-        if (!characterController.isGrounded)
-        {
-            moveDirection.y -= gravity * Time.deltaTime;
-        }
-
         // Move the controller
-        characterController.Move(moveDirection * Time.deltaTime);
+        return moveDirection;
     }
 
-    /// <summary>
-    /// Spawns a bullet from the pool at the player's position and sets the direction based on the mouse position.
-    /// Uses the bullet's ResetBullet() method.
-    /// </summary>
     [Command]
-    private void CmdShoot()
+    public void CmdMove(Vector3 direction)
     {
-        // Initially created a bullet clone and instantiated + spawned that.
-        // Switched to using a pool of pre-instantiated bullets for each player.
-        //bullet.transform.position = this.transform.position;
-        //GameObject bulletClone = Instantiate(bullet);
-        //NetworkServer.Spawn(bulletClone);
-
-        //bulletPool[currentBulletIndex].SetActive(true);
-        bulletPool[currentBulletIndex].GetComponent<Bullet>().ResetBullet(transform.position);
-        NetworkServer.Spawn(bulletPool[currentBulletIndex]);
-
-        currentBulletIndex++;
-        if(currentBulletIndex >= bulletPool.Count)
-        {
-            currentBulletIndex = 0;
-        }
+        characterController.Move(direction * Time.deltaTime);
     }
 
-    [Command]
-    public void CmdReduceLifeTotal(int i)
+    [Server]
+    public void TakeDamage(int i)
     {
         lifeTotal -= i;
     }
+
 
     private void OnGUI()
     {
@@ -125,7 +91,7 @@ public class PlayerController : NetworkBehaviour
             GUILayout.Label("Mouse position: " + mousePos);
             GUILayout.Label("World position: " + point.ToString("F3"));
             GUILayout.Label("Player position: " + this.transform.position);
-            GUILayout.Label("Current bullet index: " + currentBulletIndex);
+            //GUILayout.Label("Current bullet index: " + currentBulletIndex);
             GUILayout.EndArea();
 
             if (isGameOver)
