+++ Multiplayer Bullet Hell/Assets/Scripts/Shooting.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a8059b245dd121a4488c09954ca0f66b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
