+++ Multiplayer Bullet Hell/Logs/shadercompiler-UnityShaderCompiler.exe9.log
@@ -1,4 +1,3 @@
 Base path: 'C:/Program Files/Unity/Hub/Editor/2020.3.18f1/Editor/Data', plugins path 'C:/Program Files/Unity/Hub/Editor/2020.3.18f1/Editor/Data/PlaybackEngines'
 Cmd: initializeCompiler
 
-Cmd: shutdown
